pyproject.toml
[project]
name = "course5"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "python-telegram-bot (>=22.5,<23.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "django-timezone-field (>=7.2.1,<8.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
